move hook to utils file
